requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
email-validator==2.3.0
python-jose[cryptography]>=3.3.0
python-multipart==0.0.21
google-cloud-firestore>=2.11.0
bcrypt==4.3.0
redis==5.3.0


# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.28.1